@@ -7,18 +7,15 @@
     "serve": "rapide serve"
   },
   "dependencies": {
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
-    "@mui/material": "^5.0.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
   "devDependencies": {
-    "@types/react": "^17.0.29",
-    "@types/react-dom": "^17.0.9",
+    "@types/react": "^17.0.39",
+    "@types/react-dom": "^17.0.11",
     "rapide": "workspace:^0.3.1",
     "rapide-react": "workspace:^0.0.2",
-    "typescript": "^4.4.4"
+    "typescript": "^4.5.5"
   },
   "rapide": {
     "plugins": [
