@@ -1,6 +1,6 @@
 {
   "name": "rapide",
-  "version": "0.3.1",
+  "version": "0.3.2",
   "description": "web dev build tool",
   "author": "wkj",
   "license": "MIT",
@@ -15,16 +15,23 @@
     "client.js"
   ],
   "repository": "https://github.com/wkj9893/rapide",
+  "scripts": {
+    "lint": "deno lint src scripts test && tsc -p . --noEmit",
+    "fmt": "deno fmt src",
+    "build-client": "esbuild src/client/index.ts --bundle --minify --format=esm --outfile=client.js",
+    "build-server": "esbuild src/cli.ts --bundle --minify --outfile=cli.js --platform=node --external:esbuild",
+    "build": "pnpm build-client && pnpm build-server"
+  },
   "dependencies": {
-    "esbuild": "^0.13.8"
+    "esbuild": "^0.14.23"
   },
   "devDependencies": {
-    "@types/node": "^16.11.1",
-    "@types/ws": "^8.2.0",
-    "chokidar": "^3.5.2",
+    "@types/node": "^17.0.21",
+    "@types/ws": "^8.5.0",
+    "chokidar": "^3.5.3",
     "cjs-module-lexer": "^1.2.2",
-    "es-module-lexer": "^0.9.3",
-    "typescript": "^4.4.4",
-    "ws": "^8.2.3"
+    "es-module-lexer": "^0.10.0",
+    "typescript": "^4.5.5",
+    "ws": "^8.5.0"
   }
 }
