@@ -2,16 +2,16 @@
 import { build, createServer, preCreateServer, serve } from "./server";
 import { rootPath } from "./server/utils/path";
 import { version } from "../package.json";
-import { performance } from "perf_hooks";
-import path = require("path");
 import { cyan, lightBlue } from "./server/utils/color";
 import { address } from "./server/utils/network";
+import path from "path";
+import { performance } from "perf_hooks";
 
 const args = process.argv.slice(2);
 
 async function main() {
   if (!args[0] || args[0] === "dev") {
-    const startTime = performance.now();
+    const startTime = performance.now()
     const config = await preCreateServer();
     const server = await createServer(config);
     console.log(cyan(`\n  rapide ${version}`) + " dev server running at:\n");
@@ -24,35 +24,18 @@ async function main() {
     console.log(
       cyan(`  ready in ${(performance.now() - startTime).toFixed()}ms`),
     );
-    return;
-  }
-  if (args[0] === "build") {
+  } else if (args[0] === "build") {
     await build();
-    return;
-  }
-  if (args[0] === "serve") {
+  } else if (args[0] === "serve") {
     const p = args[1] ? args[1] : path.join(rootPath, "build");
     serve(p);
-    return;
-  }
-  if (args[0] === "--version") {
+  } else if (args[0] === "--version") {
     console.log("rapide", version);
-    return;
+  } else {
+    throw new Error(
+      `rapide:  '${args[0]}' is not a rapide command. See 'rapide --help'`,
+    );
   }
-  if (args[0] === "--help") {
-    console.log("rapide", version);
-    console.log(`Commands:
-    [root]
-    build [root]
-    serve [root]`);
-    console.log(`OPTIONS:
-    --help
-        Prints help information`);
-    return;
-  }
-  throw new Error(
-    `rapide:  '${args[0]}' is not a rapide command. See 'rapide --help'`,
-  );
 }
 
 main().catch((e) => console.error(e));
